+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v2.0.0 - 2022.08.18 17
+* use `Proxy` to handle logic
+* we can operate CSS variables like `Object` now
+* update lea
+
+
 ## v1.0.0 - 2022.08.18 15
 * tweak all files for publishing to npm
 * start use `CHANGLOG.md` since version `v1.0.0`